LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A8_T2-expected.txt

CONSOLE MESSAGE: line 87: SyntaxError: Parse error
S12.8_A8_T2

PASS Expected parsing failure

TEST COMPLETE
